api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.18
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0

